Convert ID to a number and separate out types from util
