+++ src/components/Admin/ManageSensors/AddSensorModal.tsx
@@ -26,7 +26,7 @@ import {SubmitButton} from '../ComponentUtil';
 import firebase, {firestore} from '../../../firebase';
 import {useAuth} from '../../../contexts/AuthContext';
 import axios from 'axios';
-import {LabelValue} from './Util';
+import {LabelValue} from './LabelValue';
 
 /**
  * Component to add a new sensor. Includes a button to make the modal pop up
